+++ test/test_snmp_set.sh
@@ -115,7 +115,7 @@ EOF
 fi
 
 function testinit(){
-    new "test params: -f $cfg"
+    new "test params: -s startup -f $cfg"
 
     if [ $BE -ne 0 ]; then
         # Kill old backend and start a new one