+++ test/test_snmp_system.sh
@@ -88,7 +88,7 @@ cat <<EOF > $fstate
 EOF
 
 function testinit(){
-    new "test params: -f $cfg -- -sS $fstate"
+    new "test params: -s init -f $cfg -- -sS $fstate"
     if [ $BE -ne 0 ]; then
     # Kill old backend and start a new one
     new "kill old backend"